Update tagging scheme
I've got a better idea of what we need to capture about each system, and it's more about how it's managed: do we own the code, is it outsourced service under contract, or is it outside our realm of control.

Update all components with my knowledge, and update colour scheme.